# openapi.yaml
openapi: 3.0.3
info:
  title: Random Image API
  description: 一个高性能的随机图片服务，支持多线程、Redis缓存、CORS支持和智能缓存管理
  version: 2.1.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
paths:
...
  /update:
    get:
      summary: 手动更新图片列表
      description: 手动触发图片列表更新
      operationId: updateImageList
      parameters:
        - name: token
          in: query
          description: 验证令牌
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /cache/clear:
    post:
      summary: 清空缓存
      description: 清空所有缓存
      operationId: clearCache
      parameters:
        - name: token
          in: query
          description: 验证令牌
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 缓存已清空
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /cache/reset:
    post:
      summary: 重置Redis重试计数
      description: 重置Redis重试计数
      operationId: resetRedisRetryCount
      parameters:
        - name: token
          in: query
          description: 验证令牌
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 重试计数已重置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cache/ttl/{key}:
    get:
      summary: 获取缓存TTL
      description: 获取缓存键的过期时间
      operationId: getCacheTTL
      parameters:
        - name: key
          in: path
          description: 缓存键
          required: true
          schema:
            type: string
      responses:
        '200':
          description: TTL信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  ttl:
                    type: integer
                  message:
                    type: string
                  timestamp:
                    type: string
  /cache/expire/{key}:
    post:
      summary: 设置缓存过期时间
      description: 设置缓存键的过期时间
      operationId: setCacheExpire
      parameters:
        - name: key
          in: path
          description: 缓存键
          required: true
          schema:
            type: string
      requestBody:
        description: 过期时间设置
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seconds:
                  type: integer
                  default: 3600
      responses:
        '200':
          description: 过期时间已设置
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  key:
                    type: string
                  seconds:
                    type: integer
                  message:
                    type: string
                  timestamp:
                    type: string
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 是否成功
        message:
          type: string
          description: 响应消息
        timestamp:
          type: string
          description: 时间戳
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误类型
        message:
          type: string
          description: 错误消息
        path:
          type: string
          description: 请求路径
        processingTime:
          type: integer
          description: 处理时间（毫秒）
        timestamp:
          type: string
          description: 时间戳
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: token
      description: API密钥认证
security:
  - ApiKeyAuth: []